Fixed typo

+++ src/unsw/dungeon/GameState.java
@@ -1,5 +1,5 @@
 package unsw.dungeon;
 
 public interface GameState {
-    public void run(int timeToRunFor);
+    public void run(double timeToRunFor);
 }
